Switch int to long in sales

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
 
-        int[] sales = {1, 2, 3, 4, 5, 6};
+        long[] sales = {1, 2, 3, 4, 5, 6};
         SalesManager salesManager = new SalesManager(sales);
         System.out.println("Max:" +salesManager.max());
     }
